+++ gix-features/src/parallel/in_parallel.rs
@@ -273,8 +273,6 @@ where
                                             // The first consumer to fail is the one that caused the stop - every
                                             // error thereafter is a mere reaction to the stop signal, like a
                                             // bail-out on interrupt, and must not mask the causal error.
-                                            // Causality is tracked on its own flag as failing consumers may set
-                                            // the (exposed) stop flag themselves before returning.
                                             let is_cause = !a_consumer_failed.swap(true, Ordering::SeqCst);
                                             stop_everything.store(true, Ordering::Relaxed);
                                             return Err((err, is_cause));
@@ -289,16 +287,16 @@ where
                         .expect("valid name")
                 })
                 .collect();
-            let mut cause = None;
-            let mut reaction = None;
             for thread in threads {
                 match thread.join() {
                     Ok(Ok(rval)) => {
                         results.push(rval);
                     }
                     Ok(Err((err, is_cause))) => {
-                        let slot = if is_cause { &mut cause } else { &mut reaction };
-                        slot.get_or_insert(err);
+                        // is-cause is race-free, and must be set on the first error.
+                        if is_cause {
+                            return Err(err);
+                        }
                     }
                     Err(err) => {
                         // a panic happened, stop the world gracefully (even though we panic later)
@@ -306,9 +304,6 @@ where
                         std::panic::resume_unwind(err);
                     }
                 }
-            }
-            if let Some(err) = cause.or(reaction) {
-                return Err(err);
             }
 
             stop_everything.store(true, Ordering::Relaxed);
